requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
docutils==0.16
flex==6.14.1
genson==1.2.1
idna==2.10
importlib-metadata==1.7.0
jsonpath-ng==1.5.1
jsonpointer==2.0
jsonschema==3.2.0
ply==3.11
Pygments==2.6.1
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
RESTinstance==1.0.2
rfc3987==1.3.8
robotframework==3.2.1
robotframework-requests==0.7.0
six==1.15.0
strict-rfc3339==0.7
tzlocal==2.1
urllib3==1.25.10
validate-email==1.3
zipp==3.1.0
